Move types to places that make more sense
